requirements.txt
attrs==21.4.0
bcrypt==3.2.2
boto3==1.24.21
botocore==1.27.21
cffi==1.15.1
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
Flask-WTF==1.0.1
furl==2.1.3
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
mongoengine==0.24.1
orderedmultidict==1.0.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyJWT==1.7.1
pymongo==4.1.1
pyparsing==3.0.9
pytest==7.1.2
pytest-mock==3.8.1
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
